% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_add.R
\name{ml_add}
\alias{ml_add}
\title{Add a model and training data to database}
\usage{
ml_add(model, X, Y, title, comment = "")
}
\arguments{
\item{model}{model object}

\item{X}{data.frame containing training data}

\item{Y}{vector containing response data}

\item{title}{character title of the model-data environment}

\item{comment}{a brief description associated with the commit}
}
\description{
Add a model and training data to database
}
