% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_checkout.R
\name{ml_checkout}
\alias{ml_checkout}
\title{Checkout a model and associated data files from version control}
\usage{
ml_checkout(repo, version)
}
\arguments{
\item{repo}{character string of a model repository}

\item{version}{character string of the specific version}
}
\description{
Checkout a model and associated data files from version control
}
