% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_delete.R
\name{ml_delete}
\alias{ml_delete}
\title{Delete model version control history}
\usage{
ml_delete(repo, version)
}
\arguments{
\item{repo}{character string of a model repository}

\item{version}{character string of the specific model version to be deleted. Not implemented}
}
\description{
Delete model version control history
}
